requirements.txt
kagglehub
numpy
pandas
scikit-learn
umap-learn
hdbscan
pillow
opencv-python
tqdm
torch
torchvision
open_clip_torch
moviepy
librosa
ffmpeg-python
piexif
exifread
matplotlib
midiutil
pyfluidsynth
